move paths into trees
