+++ pysyte/types/paths.py
@@ -14,236 +14,16 @@ from functools import singledispatch
 from importlib import import_module
 from typing import Iterable
 from typing import List
-from typing import Sequence
-from typing import Tuple
-from typing import Union
 
 from deprecated import deprecated
 
-from path import Path as path_Path
 from pysyte.types.lists import flatten
-from pysyte.types.methods import Method
-
-
-class PathError(Exception):
-    """Something went wrong with a path"""
-
-    prefix = "Path Error"
-
-
-class MissingPath(PathError):
-    def __init__(self, path, desc=""):
-        self.path = path
-        description = desc or "path"
-        super().__init__(f"Missing {description}{path}")
-
-
-class MissingImport(MissingPath):
-    def __init__(self, module):
-        self.module = module
-        try:
-            path_ = module.__file__
-        except AttributeError:
-            path_ = module.__name__
-        super().__init__(path_, desc="module")
-
-
-class PathAssertions:
-    """Assertions that can be made about paths"""
-
-    def assertExists(self):
-        if not self.exists():
-            raise MissingPath(self)
-        return self
-
-    def assert_isdir(self):
-        """Raise a PathError if this path is not a directory on disk"""
-        if not self.isdir():
-            raise PathError(f"{self} is not a directory")
-        return self
-
-    def assert_isfile(self):
-        """Raise a PathError if this path is not a file on disk"""
-        if not self.isfile():
-            raise PathError(f"{self} is not a file")
-        return self
-
-
-StrPath = Union["StringPath", str]  # many args can be either string or path
-
-
-class StringPath(path_Path):
-    """This class handles the path as if it were just a string
-
-    Sub-classes know about paths qua paths
-    """
-
-    # pylint: disable=abstract-method
-    # pylint: disable=too-many-public-methods
-
-    def __hash__(self):
-        return hash(str(self))
-
-    def __repr__(self) -> str:
-        string = repr(f"{self}")
-        return f"<{self.__class__.__name__} {string}>"
-
-    @deprecated(version="0.7.5", reason="Please use Python 3")
-    def __div__(self, other) -> StringPath:
-        """Need to cover parent's use of older idiom"""
-        return self.__truediv__(other)
-
-    def __truediv__(self, substring: str) -> StringPath:
-        """Handle the / operator
-
-        Add substring to self
-
-        >>> p = StringPath("/path/to")
-        >>> assert p.__truediv__("fred") == p / "fred"
-        >>> assert p / "fred" == "/path/to/fred"
-        >>> assert p / None is p
-        """
-        if not substring:
-            return self
-        full_string = os.path.join(str(self), substring)
-        return makepath(full_string)
-
-    def __floordiv__(self, substrings: Sequence[str]) -> StringPath:
-        """Handle the // operator
-
-        Add substrings to self like a path in local os
-
-        >>> p = StringPath("/path/to")
-        >>> assert p.__floordiv__("fred") == p // "fred"
-        >>> assert p // ["module", "fred.py"] == "/path/to/module/fred.py"
-        >>> assert p // None is p
-        """
-        if not substrings:
-            return self
-        string = str(self)
-        strings = [string] + list(substrings)
-        return makepath(os.path.join(*strings))
-
-    def __eq__(self, other) -> bool:
-        return str(self) == str(other)
-
-    def __lt__(self, other) -> bool:
-        return str(self) < str(other)
-
-    # functools.total_ordering does not work properly cos we inherit from str
-    # Hence: we do need to define next 3
-    def __le__(self, other):
-        return self.__eq__(other) or self.__lt__(other)
-
-    def __gt__(self, other):
-        return not self.__le__(other)
-
-    def __ge__(self, other):
-        return not self.__lt__(other)
-
-    def __contains__(self, other) -> bool:
-        """The other is in self if self.contains(other)
-
-        this method should be specialised in sub-classes
-        """
-        return self.contains(other)
-
-    def contains(self, other: StrPath) -> bool:
-        """If other is also a path then this path should start with other
-
-        E.g. /path/to/file is "in" /path
-
-        Otherwise, just use the sub-string sense of "in"
-        """
-        if isinstance(other, StringPath):
-            return str(other).startswith(str(self))
-        return str(other) in str(self)
-
-    def basename(self) -> str:
-        return str(super().basename())
-
-    @property
-    def basename_(self) -> str:
-        return self.basename()
-
-    @property
-    def name(self) -> str:
-        return str(super().name)
-
-    @property
-    def stem(self) -> StringPath:
-        stem, *_ = self.splitexts()
-        return stem
-
-    @property
-    def stem_name(self) -> str:
-        return self.stem.name
-
-    def splitexts(self) -> Tuple[StringPath, str]:
-        """Split all extensions from the path
-
-        >>> p = FilePath('here/fred.tar.gz')
-        >>> assert p.splitexts() == ('here/fred', '.tar.gz')
-        """
-        copy = self[:]
-        filename, ext = os.path.splitext(copy)
-        zippers = (
-            ".gz",
-            ".bz",
-            ".zip",
-            ".bzip",
-        )
-        for zipper in zippers:
-            if ext == zipper:
-                filename, ext_ = os.path.splitext(filename)
-                ext = f"{ext_}{zipper}"
-        return self.__class__(filename), ext
-
-    def add_ext(self, *args) -> StringPath:
-        """Join all args as extensions
-
-        Strip any leading `.` from args
-
-        >>> source = makepath(__file__)
-        >>> new = source.add_ext('txt', '.new')
-        >>> assert new.name.endswith('.py.txt.new')
-        """
-        exts = [(a[1:] if a[0] == "." else a) for a in args]
-        string = ".".join([self] + list(exts))
-        return makepath(string)
-
-    def add_missing_ext(self, ext: str) -> StringPath:
-        """Add that extension, if it is missing
-
-        >>> fred = makepath("fred")
-        >>> assert fred.add_missing_ext("") == fred
-        >>> fred_py = makepath("fred.py")
-        >>> assert fred.add_missing_ext(".py") == fred_py
-        >>> assert fred_py.add_missing_ext(".txt") == "fred.py.txt"
-        """
-        dot_ext = f'.{ext.lstrip(".")}'
-        copy = self[:]
-        _, self_ext = os.path.splitext(copy)
-        return makepath(self) if self_ext == dot_ext else self.add_ext(dot_ext)
-
-    def extend_by(self, ext: str) -> StringPath:
-        """The path to the file changed to use the given ext
-
-        >>> fred = "/path/to/fred.fred"
-        >>> assert makepath("/path/to/fred").extend_by("fred") == fred
-        >>> assert makepath("/path/to/fred.txt").extend_by(".fred") == fred
-        >>> assert makepath("/path/to/fred.txt").extend_by("..fred") == fred
-        """
-        copy = self[:]
-        filename, _ = os.path.splitext(copy)
-        ext_ = ext.lstrip(".")
-        return makepath(f"{filename}.{ext_}")
-
-    def has_vcs_dir(self):
-        for vcs_dir in (".git", ".svn", ".hg"):
-            if self.fnmatch_part(vcs_dir):
-                return True
-        return False
+from pysyte.types.trees.paths import PathPath
+from pysyte.types.trees.files import FilePath
+from pysyte.types.trees.dirs import DirectPath
+from pysyte.types.trees.strings import NoPath
+from pysyte.types.trees.errors import MissingImport
+from pysyte.types.trees.errors import PathError
 
 
 def ext_language(ext, exts=None, simple=True):
@@ -266,484 +46,6 @@ def ext_language(ext, exts=None, simple=True):
     }
     ext_languages = {_: languages[_] for _ in exts} if exts else languages
     return ext_languages.get(ext)
-
-
-class NonePath(StringPath):
-    def __init__(self, string=None):
-        self.string = string if string else ""
-        self.proxy = FilePath(self.string) or DirectPath(self.string)
-
-    def __str__(self):
-        return self.string
-
-    def __repr__(self):
-        string = self.string
-        if self.proxy:
-            string = str(self.proxy)
-        return f'<{self.__class__.__name__} "{string}">'
-
-    def __bool__(self):
-        return False
-
-    def __eq__(self, other):
-        if self.string:
-            return str(self) == str(other)
-        return not other
-
-    def __lt__(self, other):
-        if self.string and other:
-            return str(self) < str(other)
-        return bool(other)
-
-    def contains(self, other: StrPath) -> bool:
-        """As this is not a real path, just use the substring sense"""
-        return str(other) in str(self)
-
-    def __truediv__(self, child):
-        result = os.path.join(self.string, child) if child else self.string
-        return makepath(result)
-
-    @property
-    def parent(self):
-        if "/" not in self.string:
-            return None
-        parent_string = "/".join(self.string.split("/")[:-1])
-        return makepath(parent_string)
-
-    def exists(self):
-        return False
-
-    isdir = isfile = isexec = isroot = exists
-
-    def __getattr__(self, name):
-        return getattr(self.proxy, name, None)
-
-    def makedirs(self):
-        os.makedirs(str(self))
-
-
-class DotPath(StringPath):
-    """This class add path-handling to a string"""
-
-    def parent_directory(self):
-        if self.isroot():
-            return None
-        return self.parent
-
-    def parent_directories(self):
-        if self.isroot():
-            return []
-        parent = self.parent
-        return [parent] + parent.parent_directories()
-
-    def directory(self):
-        """Return a path to the path's directory"""
-        return self.parent
-
-    def dirnames(self):
-        """Split the dirname into individual directory names
-
-        An absolute path starts with an empty string, a relative path does not
-
-        >>> DotPath('/path/to/module.py').dirnames() == ['/', 'path', 'to']
-        True
-        >>> DotPath('path/to/module.py').dirnames() == ['path', 'to']
-        True
-        """
-        return [str(_) for _ in self.directory().split(os.path.sep)]
-
-    def dirpaths(self):
-        """Split the dirname into individual directory names
-
-        An absolute path starts with an empty string, a relative path does not
-
-        >>> p = DotPath('/path/to/x.py')
-        >>> assert p.paths == p.dirpaths()
-        """
-        parts = self.split()
-        result = [DotPath(parts[0] or "/")]
-        for name in parts[1:]:
-            result.append(result[-1] / name)
-        return result
-
-    def directories(self):
-        """Split the dirname into individual directory names
-
-        No empty parts are included
-
-        >>> DotPath('path/to/module.py').directories() == ['path', 'to']
-        True
-        >>> DotPath('/path/to/module.py').directories() == ['/', 'path', 'to']
-        True
-        """
-        return [d for d in self.dirnames() if d]
-
-    parents = property(
-        dirnames,
-        None,
-        None,
-        """ This path's parent directories, as a list of strings.
-
-        >>> DotPath('/path/to/module.py').parents == ['/', 'path', 'to']
-        True
-        """,
-    )
-
-    paths = property(
-        dirpaths,
-        None,
-        None,
-        """ This path's parent directories, as a sequence of paths.
-
-        >>> paths = DotPath('/usr/bin/vim').paths
-        >>> assert paths[-1].exists()  # vim might be a link
-        >>> assert paths[-2] == paths[-1].parent
-        >>> assert paths[-3] == paths[-2].parent
-        >>> assert paths[-4] == paths[-3].parent
-        >>> assert paths[-4] == paths[0]
-        """,
-    )
-
-    def path_split(self, sep=None, maxsplit=-1):
-        separator = sep or os.path.sep
-        parts = super().split(separator, maxsplit)
-        parts[0] = makepath(parts[0] if parts[0] else "/")
-        return parts
-
-    split = path_split
-
-    def abspath(self):
-        return makepath(os.path.abspath(str(self)))
-
-    def slashpath(self):
-        return self + "/" if self.isdir() else self
-
-    def short_relative_path_to(self, destination):
-        """The shorter of either the absolute path of the destination,
-            or the relative path to it
-
-        >>> print(
-        ...     DotPath('/home/guido/bin').short_relative_path_to(
-        ...         '/home/guido/build/python.tar'
-        ...     )
-        ... )
-        ../build/python.tar
-        >>> print(
-        ...     DotPath('/home/guido/bin').short_relative_path_to(
-        ...         '/mnt/guido/build/python.tar'
-        ...     )
-        ... )
-        /mnt/guido/build/python.tar
-        """
-        relative = self.relpathto(destination)
-        absolute = self.__class__(destination).abspath()
-        if len(str(relative)) < len(str(absolute)):
-            return relative
-        return absolute
-
-    def short_relative_path_to_here(self):
-        """A short path relative to current working directory"""
-        return self.short_relative_path_to(os.getcwd())
-
-    def short_relative_path_from_here(self):
-        """A short path relative to self to the current working directory"""
-        return self.__class__(os.getcwd()).short_relative_path_to(self)
-
-    def fnmatch_basename(self, glob):
-        if glob.startswith(os.path.sep):
-            glob = glob.lstrip(os.path.sep)
-        string = self.basename()
-        if fnmatch(string, glob):
-            return self
-        return None
-
-    def fnmatch_directory(self, glob):
-        if glob.startswith(os.path.sep) or glob.endswith(os.path.sep):
-            glob = glob.strip(os.path.sep)
-        if self.isdir():
-            string = self.basename()
-        else:
-            string = self.parent.basename()
-        if fnmatch(string, glob):
-            return self
-        return None
-
-    def fnmatch_directories(self, glob):
-        if glob.startswith(os.path.sep) or glob.endswith(os.path.sep):
-            glob = glob.strip(os.path.sep)
-        strings = reversed(self.directory().splitall()[1:])
-        for string in strings:
-            if fnmatch(string, glob):
-                return self
-        return None
-
-    def fnmatch_part(self, glob):
-        if self.fnmatch(glob):
-            return self
-        if self.fnmatch_basename(glob):
-            return self
-        if self.fnmatch_directory(glob):
-            return self
-        if self.fnmatch_directories(glob):
-            return self
-        return None
-
-    def expand(self):
-        """Expand the path completely
-
-        This removes any "~" (for home dir) and any shell variables
-            eliminates any symbolic links
-            and converts from relative to absolute path
-        """
-        u = os.path.expanduser(str(self))
-        v = os.path.expandvars(u)
-        r = os.path.realpath(v)
-        return makepath(r)
-
-    def same_path(self, other):
-        """Whether this path points to same place as the other"""
-        return self.expand() == other.expand()
-
-    def isroot(self):
-        raise NotImplementedError("Only a directory path can be a root")
-
-    def ishidden(self):
-        """A 'hidden file' has a name starting with a '.'"""
-        s = str(self.basename())
-        return s and s[0] == "."
-
-    def is_executable(self):
-        """Whether the path is executable"""
-        # pylint: disable=no-self-use
-        return False
-
-    def isexec(self):
-        return self.is_executable()
-
-    def has_executable(self):
-        """Whether the path has any executable bits set"""
-        executable_bits = stat.S_IEXEC | stat.S_IXGRP | stat.S_IXOTH
-        try:
-            return bool(os.stat(self).st_mode & executable_bits)
-        except OSError:
-            return False
-
-
-class FilePath(DotPath, PathAssertions):
-    """A path to a known file"""
-
-    def __truediv__(self, child):
-        raise PathError("%r has no children" % self)
-
-    def __iter__(self):
-        for line in self.stripped_lines():
-            yield line
-
-    def contains(self, other: StrPath) -> bool:
-        """Whether other is in this file's text"""
-        return str(other) in self.text()
-
-    def stripped_lines(self):
-        """A list of all lines without trailing whitespace
-
-        If lines can not be read (e.g. no such file) then an empty list
-        """
-        try:
-            return [_.rstrip() for _ in self.lines(retain=False)]
-        except (OSError, IOError, UnicodeDecodeError):
-            return []
-
-    def stripped_whole_lines(self):
-        """A list of all lines without trailing whitespace or blank lines"""
-        return [_ for _ in self.stripped_lines() if _]
-
-    def non_comment_lines(self):
-        """A list of all non-empty, non-comment lines"""
-        return [_ for _ in self.stripped_whole_lines() if not _.startswith("#")]
-
-    def isroot(self):
-        """A file cannot be root of a filesystem"""
-        return False
-
-    def is_executable(self):
-        """Whether the file has any executable bits set"""
-        return self.has_executable()
-
-    def has_line(self, string):
-        for line in self:
-            if string == line:
-                return True
-        return False
-
-    def any_line_has(self, string):
-        for line in self:
-            if string in line:
-                return True
-        return False
-
-    def as_python(self):
-        """The path to the file with a .py extension
-
-        >>> assert FilePath("/path/to/fred.txt").as_python() == "/path/to/fred.py"
-        """
-        return self.extend_by(".py")
-
-    def make_read_only(self):
-        """chmod the file permissions to -r--r--r--"""
-        self.chmod(ChmodValues.readonly_file)
-
-    def cd(self):  # pylint: disable=invalid-name
-        """Change program's current directory to self"""
-        return cd(self.parent)
-
-    def dirname(self):
-        return DirectPath(os.path.dirname(self))
-
-    parent = property(dirname)
-
-    def shebang(self):
-        """The  #! entry from the first line of the file
-
-        If no shebang is present, return an empty string
-        """
-        try:
-            first_line = self.stripped_lines()[0]
-            if first_line.startswith("#!"):
-                return first_line[2:].strip()
-        except IndexError:
-            pass
-        return ""
-
-    def mv(self, destination):  # pylint: disable=invalid-name
-        return self.move(destination)
-
-    @property
-    def language(self):
-        """The language of this file"""
-        try:
-            return self._language
-        except AttributeError:
-            self._language = ext_language(self.ext)
-        return self._language
-
-    @language.setter
-    def language(self, value):
-        self._language = value
-
-
-class DirectPath(DotPath, PathAssertions):
-    """A path which knows it might be a directory
-
-    And that files are in directories
-    """
-
-    __file_class__ = FilePath
-
-    def __iter__(self):
-        for a_path in self.listdir():
-            yield a_path
-
-    def contains(self, other: StrPath) -> bool:
-        """If other is a path then use that sense of "in"
-
-        So /path/to/here is "in" /path
-
-        Otherwise see if other is listed "in" this directory
-        """
-        if isinstance(other, DotPath):
-            return self in other.parent_directories()
-        return str(other) in [_.name for _ in self.listdir()] + [".", ".."]
-
-    def directory(self):
-        """Return a path to a directory.
-
-        Either the path itself (if it is a directory), or its parent)
-        """
-        if self.isdir():
-            return self
-        return self.parent
-
-    def remove_dir(self):
-        """Try to remove the path
-
-        If it is a directory, try recursive removal of contents too
-        """
-        if self.islink():
-            self.unlink()
-        elif self.isdir():
-            self.empty_directory()
-            if self.isdir():
-                self.rmdir()
-        else:
-            return False
-        return True
-
-    def empty_directory(self):
-        """Remove all contents of a directory
-
-        Including any sub-directories and their contents"""
-        for child in self.walkfiles():
-            child.remove()
-        for child in reversed([self.walkdirs()]):
-            if child == self or not child.isdir():
-                continue
-            child.rmdir()
-
-    def cd(self):  # pylint: disable=invalid-name
-        """Change program's current directory to self"""
-        return cd(StringPath(self))
-
-    def list_dirs(self, pattern=None):
-        return self.list_dirs_files(pattern)[0]
-
-    def list_files(self, pattern=None):
-        return self.list_dirs_files(pattern)[1]
-
-    def list_dirsfiles(self, pattern=None):
-        dirs, others = self.list_dirs_files(pattern)
-        return dirs + others
-
-    def list_dirs_files(self, pattern=None):
-        items = self.listdir(pattern)
-        dirs = [_ for _ in items if _.isdir()]
-        others = [_ for _ in items if not _.isdir()]
-        return dirs, others
-
-    def make_read_only(self):
-        """chmod the directory permissions to -r-xr-xr-x"""
-        self.chmod(ChmodValues.readonly_directory)
-
-    def touch_file(self, filename):
-        """Touch a file in the directory"""
-        path_to_file = self.__file_class__(os.path.join(self, filename))
-        path_to_file.touch()
-        return path_to_file
-
-    def existing_sub_paths(self, sub_paths):
-        """Those in the given list of sub_paths which do exist"""
-        paths_to_subs = [self / _ for _ in sub_paths]
-        return [_ for _ in paths_to_subs if _.exists()]
-
-    # pylint: disable=arguments-differ
-    def walkdirs(self, pattern=None, errors="strict", ignores=None):
-        ignored = ignore_fnmatches(ignores)
-        for path_to_dir in super(DirectPath, self).walkdirs(pattern, errors):
-            if not ignored(path_to_dir.relpath(self)):
-                yield path_to_dir
-
-    # pylint: disable=arguments-differ
-    def walkfiles(self, pattern=None, errors="strict", ignores=None):
-        ignored = ignore_fnmatches(ignores)
-        for path_to_file in super(DirectPath, self).walkfiles(pattern, errors):
-            if not ignored(path_to_file.relpath(self)):
-                yield path_to_file
-
-    def listfiles(self, pattern=None, ignores=None):
-        ignored = ignore_fnmatches(ignores)
-        return [_ for _ in self.listdir(pattern) if _.isfile() and not ignored(_)]
-
-    def isroot(self):
-        return str(self) == "/"
 
 
 def ignore_fnmatches(ignores):
@@ -778,7 +80,10 @@ def _make_module_path(arg):
 @singledispatch
 def makepath(arg) -> StringPath:
     attribute = getattr(arg, "path", "")
-    return makepath(attribute) if attribute else makepath(str(arg))
+    return makepath(attribute)
+    if attribute:
+        return attribute
+    raise NotImplementedError(f"Cannot find a path in `{arg!r}`")
 
 
 path = makepath
@@ -787,10 +92,10 @@ path = makepath
 @makepath.register(type(None))
 def _mp(arg) -> StringPath:
     """In the face of ambiguity, refuse the temptation to guess."""
-    return NonePath()
+    raise NotImplementedError(f"Zilch: {arg!r}")
 
 
-@makepath.register(DotPath)
+@makepath.register(PathPath)
 def __mp(arg) -> StringPath:
     return arg
 
@@ -816,10 +121,10 @@ def ____mp(arg) -> StringPath:
     v = os.path.expandvars(arg)
     u = os.path.expanduser(v)
     if arg == u:
-        return NonePath(arg)
+        return NoPath(arg)
     if os.path.exists(u):
         return makepath(u)
-    return NonePath(arg)
+    return NoPath(arg)
 
 
 def imports():
@@ -830,7 +135,7 @@ def imports():
 def _____mp(arg) -> StringPath:
     """Make a path from a module"""
     if arg.__name__ == "builtins":
-        return NonePath("builtins")
+        return NoPath("builtins")
     try:
         return makepath(arg.__file__)
     except AttributeError:
@@ -854,32 +159,24 @@ def _____mp(arg) -> StringPath:
 @makepath.register(type(makepath))
 def ______mp(arg) -> StringPath:
     """Make a path from a function's module"""
-    method = Method(arg)
-    stdin_regexp = re.compile("<(stdin|.*python-input.*)>")
-    if stdin_regexp.match(method.filename):
-        return NonePath(method.filename)
+    terminal_regexp = re.compile("<(stdin|.*python-input.*)>")
+    method = getattr(arg, "__wrapped__", arg)
+    filename = method.__code__.co_filename
+    if terminal_regexp.match(filename):
+        return NoPath(filename)
     return _make_module_path(method)
 
 
-@makepath.register(type(DotPath))
+@makepath.register(type(PathPath))
 def _______mp(arg) -> StringPath:
     """Make a path from a class's module"""
     return _make_module_path(arg)
 
 
-@dataclass
-class Paths:
-    """A collection of paths"""
-
-    paths: list[StrPath]
-
-    def __iter__(self):
-        yield self.paths
-
-
 @singledispatch
-def makepaths(arg: Paths) -> Paths:
-    return arg
+def makepaths(arg) -> Paths:
+    """In the face of ambiguity, refuse the temptation to guess."""
+    raise NotImplementedError(f"Do not know the type of arg: {arg!r}")
 
 
 @makepaths.register(type(None))
@@ -897,14 +194,9 @@ def ___mps(arg) -> Paths:
     return Paths([makepath(arg)])
 
 
-@makepaths.register(StrPath)
+@makepaths.register(StringPath)
 def ____mps(arg) -> Paths:
     return Paths([arg])
-
-
-@makepaths.register(str)
-def _____mps(arg) -> Paths:
-    return Paths([makepath(arg)])
 
 
 @deprecated(reason="use pathstr()", version="0.7.57")
@@ -916,7 +208,7 @@ def pathstr(string: str) -> StringPath:
     """Make a path from a string"""
     if os.path.isfile(string) or os.path.isdir(string):
         return makepath(string)
-    return NonePath(string)
+    return NoPath(string)
 
 
 def cd(path_to: StringPath) -> bool:
@@ -1044,7 +336,7 @@ def unique_first_dirs(path_strings):
     return set(first_dirs(path_strings))
 
 
-def paths_in_directory(path_: StringPath) -> List[DotPath]:
+def paths_in_directory(path_: StringPath) -> List[PathPath]:
     """Get all items in the given directory
 
     Swallow errors to give an empty list
