+++ requirements/linting.txt
@@ -1,4 +1,10 @@
 black
 blackdoc
 flake8
+flake8-bugbear
+flake8-comprehensions
+flake8-eradicate
+flake8-isort
 mypy
+types-requests
+types-mock
